requirements.txt
black==23.10.1
coloredlogs==15.0.1
Flask==2.3.3
flask-restx==1.2.0
python-dotenv==1.0.0
waitress==2.1.2